provisions & storeProvisions modules

+++ backend/internal/modules/product/product.handler.go
@@ -53,7 +53,7 @@ func (h *ProductHandler) GetProductDetails(c *gin.Context) {
 		return
 	}
 
-	productDetails, err := h.service.GetProductByID(uint(productID))
+	productDetails, err := h.service.GetProductByID(productID)
 	if err != nil {
 		switch {
 		case errors.Is(err, types.ErrProductNotFound):
@@ -123,7 +123,7 @@ func (h *ProductHandler) GetProductSizesByProductID(c *gin.Context) {
 		return
 	}
 
-	productSizes, err := h.service.GetProductSizesByProductID(uint(productID))
+	productSizes, err := h.service.GetProductSizesByProductID(productID)
 	if err != nil {
 		localization.SendLocalizedResponseWithKey(c, types.Response500ProductSizeGet)
 		return
@@ -139,7 +139,7 @@ func (h *ProductHandler) GetProductSizeByID(c *gin.Context) {
 		return
 	}
 
-	productSize, err := h.service.GetProductSizeDetailsByID(uint(productSizeID))
+	productSize, err := h.service.GetProductSizeDetailsByID(productSizeID)
 	if err != nil {
 		switch {
 		case errors.Is(err, types.ErrProductSizeNotFound):
@@ -211,7 +211,7 @@ func (h *ProductHandler) UpdateProduct(c *gin.Context) {
 		return
 	}
 
-	existingProduct, err := h.service.UpdateProduct(uint(productID), &dto)
+	existingProduct, err := h.service.UpdateProduct(productID, &dto)
 	if err != nil {
 		localization.SendLocalizedResponseWithKey(c, types.Response500ProductUpdate)
 		return
@@ -219,7 +219,7 @@ func (h *ProductHandler) UpdateProduct(c *gin.Context) {
 
 	action := types.UpdateProductAuditFactory(
 		&data.BaseDetails{
-			ID:   uint(productID),
+			ID:   productID,
 			Name: existingProduct.Name,
 		},
 		&dto,
@@ -245,13 +245,13 @@ func (h *ProductHandler) UpdateProductSize(c *gin.Context) {
 		return
 	}
 
-	existingProductSize, err := h.service.GetProductSizeDetailsByID(uint(productSizeID))
+	existingProductSize, err := h.service.GetProductSizeDetailsByID(productSizeID)
 	if err != nil {
 		localization.SendLocalizedResponseWithKey(c, types.Response500ProductSizeUpdate)
 		return
 	}
 
-	err = h.service.UpdateProductSize(uint(productSizeID), input)
+	err = h.service.UpdateProductSize(productSizeID, input)
 	if err != nil {
 		localization.SendLocalizedResponseWithKey(c, types.Response500ProductSizeUpdate)
 		return
@@ -259,7 +259,7 @@ func (h *ProductHandler) UpdateProductSize(c *gin.Context) {
 
 	action := types.UpdateProductSizeAuditFactory(
 		&data.BaseDetails{
-			ID:   uint(productSizeID),
+			ID:   productSizeID,
 			Name: existingProductSize.Name,
 		},
 		input,
@@ -279,7 +279,7 @@ func (h *ProductHandler) DeleteProduct(c *gin.Context) {
 		return
 	}
 
-	existingProduct, err := h.service.DeleteProduct(uint(productID))
+	existingProduct, err := h.service.DeleteProduct(productID)
 	if err != nil {
 		localization.SendLocalizedResponseWithKey(c, types.Response500ProductDelete)
 		return
@@ -287,7 +287,7 @@ func (h *ProductHandler) DeleteProduct(c *gin.Context) {
 
 	action := types.DeleteProductAuditFactory(
 		&data.BaseDetails{
-			ID:   uint(productID),
+			ID:   productID,
 			Name: existingProduct.Name,
 		},
 	)
@@ -306,13 +306,13 @@ func (h *ProductHandler) DeleteProductSize(c *gin.Context) {
 		return
 	}
 
-	existingProduct, err := h.service.GetProductByID(uint(productSizeID))
+	existingProduct, err := h.service.GetProductByID(productSizeID)
 	if err != nil {
 		localization.SendLocalizedResponseWithKey(c, types.Response500ProductSizeDelete)
 		return
 	}
 
-	err = h.service.DeleteProductSize(uint(productSizeID))
+	err = h.service.DeleteProductSize(productSizeID)
 	if err != nil {
 		localization.SendLocalizedResponseWithKey(c, types.Response500ProductSizeDelete)
 		return
@@ -320,7 +320,7 @@ func (h *ProductHandler) DeleteProductSize(c *gin.Context) {
 
 	action := types.DeleteProductSizeAuditFactory(
 		&data.BaseDetails{
-			ID:   uint(productSizeID),
+			ID:   productSizeID,
 			Name: existingProduct.Name,
 		},
 	)
